Split reusable_message_sender.py file in message_sender package
