+++ bot/logger/message_sender/__init__.py
@@ -1,0 +1,3 @@
+class MessageSender:
+    def send(self, text):
+        raise NotImplementedError()
